fixed linting issues and added bazel
