+++ .bazelrc
@@ -1,0 +1,3 @@
+build --@io_bazel_rules_go//go/config:pure
+test --test_output=errors
+build --symlink_prefix=dist/
