+++ .gitignore
@@ -17,3 +17,4 @@ tmp
 # Dependency directories (remove the comment below to include it)
 vendor
 bin
+dist
