--- components/hrple-cli/scripts/docker-release.sh
@@ -1,11 +1,0 @@
-#!/usr/bin/env bash
-
-source $(dirname "$0")/common.sh
-getProjectDir
-
-docker build . --rm --force-rm --target release -t hrple:latest --build-arg RELEASE=0 --build-arg COMPONENT=hrple-cli --build-arg COMPONENT_EP=/hrple/hrple-cli
-
-echo
-echo -e "==> ${YELLOW}Results:${NC}"
-docker image ls -f label="app=hrple"
-echo
