+++ deps.bzl
@@ -1,0 +1,2 @@
+def go_dependencies():
+    pass
