--- go.mod
@@ -1,3 +1,0 @@
-module github.com/brettmostert/hrple
-
-go 1.18
