+++ go/components/hrple-cli/cmd/hrple-cli/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "hrple-cli_lib",
+    srcs = ["main.go"],
+    importpath = "github.com/brettmostert/hrple/go/components/hrple-cli/cmd/hrple-cli",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//go/components/hrple-cli/internal/commands",
+        "//go/pkg/errors/exitError",
+    ],
+)
+
+go_binary(
+    name = "hrple-cli",
+    embed = [":hrple-cli_lib"],
+    visibility = ["//visibility:public"],
+)
