+++ go/components/hrple-cli/internal/commands/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "commands",
+    srcs = [
+        "add.go",
+        "executor.go",
+        "test.go",
+    ],
+    importpath = "github.com/brettmostert/hrple/go/components/hrple-cli/internal/commands",
+    visibility = ["//go/components/hrple-cli:__subpackages__"],
+    deps = [
+        "//go/pkg/cli",
+        "//go/pkg/errors/exitError",
+    ],
+)
