+++ go/components/hrple-cli/scripts/docker-build.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+# # Build a image with src code and debug
+# docker build --rm --force-rm --target build -t hrple:latest-dev -f ./docker/Dockerfile .
+
+# # Done!
+# echo
+# echo -e "==> ${YELLOW}Results:${NC}"
+# docker image ls -f label="app=hrple"
+# echo
