+++ go/go.mod
@@ -1,0 +1,5 @@
+module github.com/brettmostert/hrple/go
+
+replace github.com/brettmostert/hrple/go => ../
+
+go 1.18
