+++ go/pkg/errors/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "errors",
+    srcs = [
+        "error.go",
+        "reason.go",
+    ],
+    importpath = "github.com/brettmostert/hrple/go/pkg/errors",
+    visibility = ["//visibility:public"],
+)
