+++ go/pkg/errors/exitError/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "exitError",
+    srcs = [
+        "error.go",
+        "exitcodes.go",
+    ],
+    importpath = "github.com/brettmostert/hrple/go/pkg/errors/exitError",
+    visibility = ["//visibility:public"],
+)
