+++ go/pkg/testHelper/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "testHelper",
+    srcs = ["util.go"],
+    importpath = "github.com/brettmostert/hrple/go/pkg/testHelper",
+    visibility = ["//visibility:public"],
+)
